CAMEL-24000: camel-spring-boot - fail fast on duration overflow instead of silent int wraparound
Address review feedback: use Math.toIntExact instead of a narrowing cast
when converting the startup condition interval/timeout Durations to millis.

+++ core/camel-spring-boot/src/main/java/org/apache/camel/spring/boot/CamelAutoConfiguration.java
@@ -482,8 +482,8 @@ public class CamelAutoConfiguration {
     StartupConditionStrategy startupConditionStrategy(CamelStartupConditionConfigurationProperties config) {
         StartupConditionStrategy scs = new DefaultStartupConditionStrategy();
         scs.setEnabled(config.isEnabled());
-        scs.setInterval((int) config.getInterval().toMillis());
-        scs.setTimeout((int) config.getTimeout().toMillis());
+        scs.setInterval(Math.toIntExact(config.getInterval().toMillis()));
+        scs.setTimeout(Math.toIntExact(config.getTimeout().toMillis()));
         scs.setOnTimeout(config.getOnTimeout());
         String envExist = config.getEnvironmentVariableExists();
         if (envExist != null) {
